Update logic that determines project version using git (#42)
If the version includes a commit hash, limit it to the first 7
characters. Ensures that the output for a given commit is
deterministic for a given project -- the previous behavior would vary
the length to ensure that the commit hash could uniquely identify a
commit, but this determination could vary based on the local project
environment.