Allow scripts to modify os-arch-bin dists (#137)
Previously, the os-arch-bin dister would only add the single binary file
to the distribution archive, making it impossible for scripts to modify
the format of the distribution. Now the archives contain all files that
exist inside the arch directory in the working directory. This leads to
the same archives in the standard case, but will also include any script
modifications.