Add "filter-artifact-names" and "exclude-artifact-names" to publishers
Adds flags to the built-in publishers that enables controlling the behavior
of the artifacts that will be published by the operation.

If "--filter-artifact-names" is specified, then only artifacts with names
that match the specified regular expression are published.

If "--exclude-artifact-names" is specified, then any artifacts with names
that match the specified regular expression are not published.