Add Docker output directory as a concept
Add an "output-dir" concept for DockerConfig that is analgous
to what exists for BuildConfig and DistConfig. Also updates the
existing OCI output to use this concept.

This makes the concept of Docker build/dist outputs more first-class.
The removes some of the hackiness of the previous OCI implementation
and also allows for supporting Docker metadata artifacts.