Pass the --platform arg with docker output type (#586)
For convenience, we should allow cross platform build for docker output type as well and for this we need to pass the `--platform` arg.

If multiple platforms are specified, only the native platform will be picked. If one platform is specified, this platform will be picked even for cross platform.