Update distertester
Updates TestCase and RunAssetDistTest to support more robust
testing of disters.

The primary changes are the addition of the "PostTagAction" and
"ValidateOutput" fields to the TestCase struct.

If "PostTagAction" is specified, it is run after the
tester commits its files and creates the git tag.

If "ValidateOuput" is specified, it is run with the output
to provide validation beyond just equality checks.

Both fields are additive and should not change the behavior
of existing tests that do not specify the fields.